Combine OCR functionality into a single module #236
